feat(useKeyboardShortcuts): config.json adds to shortcuts
